Reject incomplete HIL captures and preserve bounded serial diagnostics

+++ examples/01_basic_bringup_cli/main.cpp
@@ -414,7 +414,7 @@ void printDriverHealth() {
   }
   const uint32_t totalOk = settings.totalSuccess;
   const uint32_t totalFail = settings.totalFailures;
-  const uint32_t total = totalOk + totalFail;
+  const uint64_t total = static_cast<uint64_t>(totalOk) + totalFail;
   const float successRate = (total > 0U)
                                 ? (100.0f * static_cast<float>(totalOk) / static_cast<float>(total))
                                 : 0.0f;
